// src/enums/CollectionName.enum.ts
export enum ChangeCollectionNameEnum {
  FAILED_TO_CHANGE_COLLECTION_NAME = "FAILED_TO_CHANGE_COLLECTION_NAME",
  COLLECTION_NAME_CHANGED = "COLLECTION_NAME_CHANGED",
}

export enum GetCollectionNamesEnum {
  GOT_COLLECTION_NAMES = "GOT_COLLECTION_NAMES",
  FAILED_TO_GET_COLLECTION_NAMES = "FAILED_TO_GET_COLLECTION_NAMES",
}

export enum GetCollectionNamesImagesIdsEnum {
  COLLECTION_NAMES_IDS = "COLLECTION_NAMES_IDS",
  FAILED_TO_RETRIVE_COLLECTION_NAME_IDS = "FAILED_TO_RETRIVE_COLLECTION_NAME_IDS",
}

export enum GetCollectionNameByIdEnum {
  FAILED_TO_RETRIVE_COLLECTION_NAME_BY_ID = "FAILED_TO_RETRIVE_COLLECTION_NAME_BY_ID",
  COLLECTION_NAMES_RETRIVED = "COLLECTION_NAMES_RETRIVED",
  COLLECTION_NOT_FOUND = "COLLECTION_NOT_FOUND",
}

export enum DeleteCollectionNameEnum {
  FAILED_TO_DELETE_COLLECTION_NAME = "FAILED_TO_DELETE_COLLECTION_NAME",
  COLLECTION_NAME_DELETED = "COLLECTION_NAME_DELETED",
}